ST-855 Refactoring to rename and attempting to add List of Bundles to CaseData
