+++ src/main/java/uk/gov/hmcts/sptribs/document/bundling/BundleDocument.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.sptribs.document.dto;
+package uk.gov.hmcts.sptribs.document.bundling;
 
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.AllArgsConstructor;
@@ -6,17 +6,15 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
-import lombok.extern.slf4j.Slf4j;
 import uk.gov.hmcts.sptribs.document.model.CICDocument;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
 @Builder
-@Slf4j
 @Getter
 @Setter
 @AllArgsConstructor
 @NoArgsConstructor
-public class CicBundleDocumentDTO {
+public class BundleDocument {
 
     private String name;
     private String description;
