+++ src/test/java/uk/gov/hmcts/sptribs/document/bundling/CicBundleDtoTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.sptribs.document.dto;
+package uk.gov.hmcts.sptribs.document.bundling;
 
 import com.fasterxml.jackson.databind.JavaType;
 import com.fasterxml.jackson.databind.JsonNode;
@@ -19,7 +19,7 @@ import static org.junit.Assert.assertNull;
 public class CicBundleDtoTest {
     private final ObjectMapper mapper = new ObjectMapper();
     private final File jsonFile = new File(ClassLoader.getSystemResource("case.json").getPath());
-    private final JavaType type = mapper.getTypeFactory().constructParametricType(CicValue.class, CicBundleDTO.class);
+    private final JavaType type = mapper.getTypeFactory().constructParametricType(CicValue.class, Bundle.class);
 
     @Test
     public void testDeserialization() throws IOException {
@@ -27,7 +27,7 @@ public class CicBundleDtoTest {
         ArrayNode bundles = (ArrayNode) root.path("case_details").path("case_data").path("caseBundles");
         JsonNode firstBundle = bundles.get(0);
 
-        CicValue<CicBundleDTO> bundleDTO = mapper.readValue(mapper.treeAsTokens(firstBundle), type);
+        CicValue<Bundle> bundleDTO = mapper.readValue(mapper.treeAsTokens(firstBundle), type);
 
         assertEquals("Bundle Title", bundleDTO.getValue().getTitle());
         assertNull(bundleDTO.getValue().getDescription());
@@ -39,13 +39,13 @@ public class CicBundleDtoTest {
         assertNotNull(bundleDTO.getValue().getPageNumberFormat());
         assertNotNull(bundleDTO.getValue().getPaginationStyle());
         assertNull(bundleDTO.getValue().getStitchedDocument());
-        CicValue<CicBundleDocumentDTO> bundleDocumentDTO = bundleDTO.getValue().getDocuments().get(0);
+        CicValue<BundleDocument> bundleDocumentDTO = bundleDTO.getValue().getDocuments().get(0);
         assertEquals("firstBundle", bundleDocumentDTO.getValue().getName());
         assertEquals("First description", bundleDocumentDTO.getValue().getDescription());
         assertEquals(1, bundleDocumentDTO.getValue().getSortIndex());
         assertNull(bundleDocumentDTO.getValue().getSourceDocument());
 
-        CicValue<CicBundleFolderDTO> bundleFolderDTO = bundleDTO.getValue().getFolders().get(0);
+        CicValue<BundleFolder> bundleFolderDTO = bundleDTO.getValue().getFolders().get(0);
         assertEquals("firstFolder", bundleFolderDTO.getValue().getName());
         assertNull(bundleFolderDTO.getValue().getFolders());
         assertEquals(1, bundleDocumentDTO.getValue().getSortIndex());
